[REFACTOR] application, register Api Domain Service 분리
